PythonScript/requirements.txt
google-generativeai
python-dotenv
pathlib
youtube_transcript_api
Flask
flask-cors